requirements.txt
Flask==3.0.2
google-generativeai==0.7.2
openai==1.35.10
anthropic==0.29.0
httpx==0.27.0
authlib==1.3.0
flask-login==0.6.3
python-dotenv==1.0.1 